refactor base dataset reader
