+++ django_project/gap_api/tests/test_measurement_api.py
@@ -55,7 +55,6 @@ class HistoricalAPITest(CommonMeasurementAPITest):
         self.assertEqual(response.data, {})
 
 
-
 class ForecastAPITest(CommonMeasurementAPITest):
     """Forecast api test case."""
